# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Professors API",
    "version": "1.0.0",
    "description": "API for managing professors"
  },
  "paths": {
    "/professors/{p_uni}": {
      "get": {
        "summary": "Get a professor by p_uni",
        "operationId": "getProfessor",
        "parameters": [
          {
            "name": "p_uni",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Professor found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Professor"
                }
              }
            }
          },
          "404": {
            "description": "Professor not found"
          }
        }
      },
      "patch": {
        "summary": "Update a professor by p_uni",
        "operationId": "updateProfessor",
        "parameters": [
          {
            "name": "p_uni",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Professor"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Professor updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Professor"
                }
              }
            }
          },
          "404": {
            "description": "Professor not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a professor by p_uni",
        "operationId": "deleteProfessor",
        "parameters": [
          {
            "name": "p_uni",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Professor deleted"
          },
          "404": {
            "description": "Professor not found"
          }
        }
      }
    },
    "/professors": {
      "get": {
        "summary": "Get all professors",
        "operationId": "getAllProfessors",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "name": "professor_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of professors",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Professor"
                      }
                    },
                    "total": {
                      "type": "integer"
                    },
                    "page": {"type": "integer"
                    },
                    "size": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new professor",
        "operationId": "createProfessor",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Professor"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Professor created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Professor"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Professor": {
        "type": "object",
        "properties": {
          "p_uni": {
            "type": "string"
          },
          "first_name": {
            "type": "string"
          },
          "last_name": {
            "type": "string"
          }
        },
        "required": ["p_uni", "first_name", "last_name"],
        "example": {
          "p_uni": "Cheese123",
          "first_name": "John",
          "last_name": "Nutz"
        }
      }
    }
  }
}